Changes on task2 and task3
