+++ task2/fileoutput/FileOutput.java
@@ -3,11 +3,12 @@ package task2.fileoutput;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
+import java.util.Scanner;
 
 /**
  * Created by clouway on 12.05.16.
  */
-public class Writer extends Reader {
+public class FileOutput {
     private String string;
     private File file;
 
@@ -15,8 +16,28 @@ public class Writer extends Reader {
      * Constructor
      * @param name
      */
-    public Writer(String name) {
+    public FileOutput(String name) {
         this.string = name;
+    }
+
+    /**
+     * Reads from the console.
+     * @param end
+     * @return
+     */
+    public String read(String end) {
+        Scanner scan = new Scanner(System.in);
+        String string = "";
+        boolean bool = true;
+        while (bool != false) {
+            String read = scan.nextLine();
+            if (read.equals(end)) {
+                bool = false;
+            } else {
+                string += read;
+            }
+        }
+        return string;
     }
 
     /**
